# target/classes/openapi.yaml
openapi: 3.0.3
info:
  title: DigiShopKart
  description: Apis of E-commerce application
  contact:
    name: Ajay Kate
    url: http://digishopkart.com
  version: '0.1'
servers:
  - url: http://localhost:8085
#------------------------------PATHS ---------------------------------
paths:
...
  /api/v1/products:
    description: create Product
    post:
      tags:
        - product
      operationId: createProduct
      summary: insert a Product
      requestBody:
        $ref: '#/components/requestBodies/ProductBody'
      responses:
        200:
          description: sucess response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        404:
          $ref: '#/components/responses/NotFound'
        400:
          $ref: '#/components/responses/BadRequest'
    get:
      tags:
        - product
      operationId: fetchaAllProducts
      summary: fetch all products
      description: all Products
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        404:
          $ref: '#/components/responses/NotFound'
        401:
          $ref: '#/components/responses/Unauthorized'
  /api/v1/products/{productId}:
    get:
      tags:
        - product
      operationId: fetchProduct
      summary: fetch a product
      description: fetch Product by id
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        404:
          $ref: '#/components/responses/NotFound'
        401:
          $ref: '#/components/responses/Unauthorized'

    delete:
      tags:
        - product
      operationId: deleteProduct
      summary: delete a product
      description: delete a product by id
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        404:
          $ref: '#/components/responses/NotFound'
        401:
          $ref: '#/components/responses/Unauthorized'

    put:
      tags:
        - product
      operationId: updateProduct
      summary: update a product
      description: updating a product by productId
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: integer
      requestBody:
        $ref: '#/components/requestBodies/ProductBody'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        404:
          $ref: '#/components/responses/NotFound'
        401:
          $ref: '#/components/responses/Unauthorized'

  #  ****************************** VARIANT  *****************************
...
  /api/v1/subscriptions:
    post:
      tags:
        - Subscription
      summary: create a suscription
      operationId: createSubscription
      description: Api to create Subscription by using Customer, Product, Discount
      requestBody:
        $ref: '#/components/requestBodies/SubscriptionBody'
      responses:
        200:
          description: Success Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        404:
          $ref: '#/components/responses/NotFound'
        401:
          $ref: '#/components/responses/Unauthorized'
    get:
      tags:
        - Subscription
      operationId: fetchAllSubscriptions
      summary: fetch all suscriptions
      description: APi to get All the sunscriptions
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        404:
          $ref: '#/components/responses/NotFound'
        401:
          $ref: '#/components/responses/Unauthorized'
  /api/v1/subscription/{subscriptionId}:
    get:
      tags:
        - Subscription
      operationId: fetchSubscription
      summary: fetch a suscription
      description: Api to fetch sunscription by id
      parameters:
        - in: path
          name: subscriptionId
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        404:
          $ref: '#/components/responses/NotFound'
        401:
          $ref: '#/components/responses/Unauthorized'

    delete:
      tags:
        - Subscription
      operationId: deleteSubscription
      summary: delete a suscription
      description: Api to delete the subscription by id
      parameters:
        - in: path
          name: subscriptionId
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        404:
          $ref: '#/components/responses/NotFound'
        401:
          $ref: '#/components/responses/Unauthorized'

    put:
      tags:
        - Subscription
      operationId: updateSubscription
      summary: update a suscription
      description: Api to update subscription record by id
      parameters:
        - in: path
          name: subscriptionId
          required: true
          schema:
            type: integer
      requestBody:
        $ref: '#/components/requestBodies/SubscriptionBody'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        404:
          $ref: '#/components/responses/NotFound'
        401:
          $ref: '#/components/responses/Unauthorized'

  #  ********************************** ORDER  **********************************
...
components:
  responses:
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errors'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errors'
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errors'
  #___________________REQUEST BODY_____________________________________
  requestBodies:
    SubscriptionBody:
      description: A JSON object Contenting Subscription information.
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Subscription'
...
    ProductBody:
      description: A JSON object Contenting Product and variant information
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Product'
...
  schemas:
    Errors:
      type: object
      properties:
        code:
          type: string
        message:
          title: string
      required:
        - code
        - message
    Customer:
      type: object
      required:
        - firstName
        - lastName
        - email
        - mobile
        - customerStatus
        - customerAddress
      properties:
        customerId:
          type: integer
        firstName:
          type: string
          example: Ajay
          pattern: '^[a-zA-Z]+(\s[a-zA-Z]+)?$'
        lastName:
          type: string
          example: Kate
          pattern: '^[a-zA-Z]+(\s[a-zA-Z]+)?$'
        email:
          type: string
          example: ajaykate@gmail.com
          format: email
        mobile:
          type: string
          example: +919876543210
        customerStatus:
          type: string
          enum:
            - active
            - inactive
          default: active
          nullable: false
        customerAddress:
          type: array
          items:
            $ref: '#/components/schemas/CustomerAddress'
    CustomerAddress:
      type: object
      required:
        - customerFullName
        - country
        - state
        - town
        - area
        - houseOrBuilding
        - landmark
        - pinCode
        - mobile
      properties:
        id:
          type: integer
        addressType:
          type: string
          enum:
            - shipping
            - billing
        customerFullName:
          type: string
        country:
          type: string
        state:
          type: string
        town:
          type: string
        area:
          type: string
        houseOrBuilding:
          type: string
        landmark:
          type: string
        pinCode:
          type: integer
        mobile:
          type: string
    Product:
      type: object
      required:
        - productName
        - createdAt
        - productCode
        - productPrice
        - productStatus
        - productCategory
      properties:
        productId:
          type: integer
        productName:
          type: string
        productCode:
          type: string
        productPrice:
          type: number
          minimum: 1
          format: double
        productCategory:
          type: string
          enum:
            - electronics
            - cloths
            - footwares
            - jwelleries
            - gadgets
        productStatus:
          type: string
          enum:
            - active
            - inactive
        variant:
          type: array
          items:
            $ref: '#/components/schemas/Variant'
    Variant:
      type: object
      required:
        - colour
        - size
      properties:
        id:
          type: integer
        name:
          type: string
          example: Size, Colour, etc...
        value:
          type: string
        status :
          type: string
          enum:
            - active
            - inactive
    Discount:
      type: object
      required:
        - discountType
        - couponValue
      properties:
        discountId:
          type: integer
        couponName:
          type: string
        discountType:
          type: string
          enum:
            - FLAT
            - PERCENTAGE
        couponValue:
          type: number
          format: double
    Subscription:
      type: object
      properties:
        subscriptionId:
          type: integer
        activatedAt:
          type: string
          format: date-time
          example: '1955-04-23T13:22:52.685Z'
        status:
          type: string
          enum: [active, inactive]
        customer:
          $ref: '#/components/schemas/Customer'
        products:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        discount:
          $ref: '#/components/schemas/Discount'
...